refactor the scraper
